Deprecate implicit use of `LoggerFactory`
