+++ slf4j/src/main/scala/org/typelevel/log4cats/slf4j/package.scala
@@ -17,20 +17,10 @@
 package org.typelevel.log4cats
 
 import cats.effect.Sync
-import org.slf4j.{Logger => JLogger}
 
 package object slf4j {
-  implicit def loggerFactoryforSync[F[_]: Sync]: Slf4jFactory[F] = new Slf4jFactory[F] {
-    override def getLoggerFromName(name: String): SelfAwareStructuredLogger[F] =
-      Slf4jLogger.getLoggerFromName(name)
 
-    override def getLoggerFromSlf4j(logger: JLogger): SelfAwareStructuredLogger[F] =
-      Slf4jLogger.getLoggerFromSlf4j(logger)
-
-    override def fromName(name: String): F[SelfAwareStructuredLogger[F]] =
-      Slf4jLogger.fromName(name)
-
-    override def fromSlf4j(logger: JLogger): F[SelfAwareStructuredLogger[F]] =
-      Slf4jLogger.fromSlf4j(logger)
-  }
+  @deprecated("LoggerFactories should be passed explicitly, not implicitly", "2.5.0")
+  implicit def loggerFactoryforSync[F[_]](implicit F: Sync[F]): Slf4jFactory[F] =
+    Slf4jFactory.create(F)
 }
